Exit on unrecoverable errors in `dbLoop`
This patch makes the `.catch()` on `dbLoop()` exit with a non-zero exit
code if an error occurs. Previously, the error was logged then ignored,
which stops new recently played songs from being ingested, which isn't
desirable, as the server keeps running but is effectively "dead".

With this change, uncaught errors in `dbLoop` (such as the db connection
dying for unknown reasons) cause the server to crash instead of dropping
into an idle state. This allows the container orchestrator running the
server to attempt and restart the server, which would allow it to come
back up when the issue is resolved. In particular, this is very useful
for gracefully handling temporary Mongo downtimes, e.g. upgrades.