Display dedicated error page if API_ENDPOINT is set to the frontend
Sometimes users set the API_ENDPOINT variable to the frontend instead of
the backend by accident. When they do that, they currently see an empty
page when they try to log in.

This commit adds a dedicated error page when the path /oauth/spotify is
accessed in the frontend. This should never happen, the corresponding
request should end up at the backend instead. Displaying a detailed
error page in that case should reduce user frustration and error
reports resulting from this misconfiguration.