Adjust documentation to discourage manually setting CORS
On the issue tracker, lots of users are struggling with CORS settings.
Looking at the issues, it's mostly because the setting is not correctly
set to the frontend domain. 99.9% of users will only ever use the
backend from the frontend domain. The server already sets the CORS
setting automatically to the frontend domain when it is not set.

This commit also removes the
i-want-a-security-vulnerability-and-want-to-allow-all-origins
configuration value from the README, as there is almost no use case for
it, and certainly not in production.

Hopefully this reduces the confusion around the CORS settings.